+++ src/Middleware/NotFoundHandler.php
@@ -10,20 +10,22 @@ declare(strict_types=1);
 
 namespace Laminas\Stratigility\Middleware;
 
-use Fig\Http\Message\StatusCodeInterface as StatusCode;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
+use Laminas\Stratigility\Handler\NotFoundHandler as NotFoundRequestHandler;
 
-use function sprintf;
-
+/**
+ * @deprecated Will be removed in v4 in favor of {@see \Laminas\Stratigility\Handler\NotFoundHandler}
+ */
 final class NotFoundHandler implements MiddlewareInterface
 {
+
     /**
-     * @var callable
+     * @var NotFoundRequestHandler
      */
-    private $responseFactory;
+    private $notFoundHandler;
 
     /**
      * @param callable $responseFactory A factory capable of returning an
@@ -32,23 +34,14 @@ final class NotFoundHandler implements MiddlewareInterface
      */
     public function __construct(callable $responseFactory)
     {
-        $this->responseFactory = function () use ($responseFactory) : ResponseInterface {
-            return $responseFactory();
-        };
+        $this->notFoundHandler = new NotFoundRequestHandler($responseFactory);
     }
 
     /**
-     * Creates and returns a 404 response.
+     * Uses the {@see \Laminas\Stratigility\Handler\NotFoundHandler} to create a 404 response.
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler) : ResponseInterface
     {
-        $response = ($this->responseFactory)()
-            ->withStatus(StatusCode::STATUS_NOT_FOUND);
-        $response->getBody()->write(sprintf(
-            'Cannot %s %s',
-            $request->getMethod(),
-            (string) $request->getUri()
-        ));
-        return $response;
+        return $this->notFoundHandler->handle($request);
     }
 }